+++ jdbc/src/main/java/net/snowflake/client/internal/api/implementation/parameters/Property.java
@@ -1,0 +1,6 @@
+package net.snowflake.client.internal.api.implementation.parameters;
+
+public interface Property {
+
+  String getKey();
+}